refactor(nika-display): the render goldens move out at the file wall
The paused card + blocked meter additions pushed render.rs to 1530
lines — over the 1500 hard cap the pre-push gate enforces. The test
module (820 lines of goldens) moves to render_tests.rs via the
path-module idiom (the run/tests.rs precedent): same module, same
super::*, zero behavior change. 711 + 825 lines, both under the wall.
Display lib 106 green.
